feat: migrate server command
